Move hub templates enum
