+++ src/main/app/case/definition.ts
@@ -1,6 +1,6 @@
 /* tslint:disable */
 /* eslint-disable */
-// Generated using typescript-generator version 2.36.1070 on 2022-08-18 09:16:02.
+// Generated using typescript-generator version 3.0.1157 on 2022-12-01 15:37:33.
 
 export interface Address {
   AddressLine1: string;
@@ -1891,34 +1891,6 @@ export const enum State {
   Submitted = 'Submitted',
   WelshTranslationRequested = 'WelshTranslationRequested',
   WelshTranslationReview = 'WelshTranslationReview',
-}
-
-export const enum HubTemplate {
-  AosAwaitingOrDrafted = '/aos-awaiting-or-drafted.njk',
-  AoSDue = '/aos-due.njk',
-  ApplicantNotYetAppliedForConditionalOrder = '/applicant-not-yet-applied-for-conditional-order.njk',
-  AwaitingAmendedApplication = '/awaiting-amended-application.njk',
-  AwaitingAoS = '/awaiting-aos.njk',
-  AwaitingBailiffService = '/awaiting-bailiff-service.njk',
-  AwaitingClarification = '/awaiting-clarification.njk',
-  AwaitingConditionalOrder = '/awaiting-conditional-order.njk',
-  AwaitingFinalOrder = '/awaiting-final-order.njk',
-  AwaitingFinalOrderOrFinalOrderOverdue = '/awaiting-final-order-or-final-order-overdue.njk',
-  AwaitingGeneralConsideration = '/awaiting-general-consideration.njk',
-  AwaitingJointFinalOrder = '/awaiting-joint-final-order.njk',
-  AwaitingLegalAdvisorReferral = '/awaiting-legal-advisor-referral.njk',
-  AwaitingLegalAdvisorReferralOrAwaitingPronouncement = '/awaiting-legal-advisor-referral-or-awaiting-pronouncement.njk',
-  AwaitingPronouncement = '/awaiting-pronouncement.njk',
-  AwaitingServiceConsiderationOrAwaitingBailiffReferral = '/awaiting-service-consideration-or-awaiting-bailiff-referral.njk',
-  BailiffServiceUnsuccessful = '/bailiff-service-unsuccessful.njk',
-  ClarificationSubmitted = '/clarification-submitted.njk',
-  ConditionalOrderPending = '/conditional-order-pending.njk',
-  ConditionalOrderPronounced = '/conditional-order-pronounced.njk',
-  FinalOrderComplete = '/final-order-complete.njk',
-  FinalOrderRequested = '/final-order-requested.njk',
-  Holding = '/holding.njk',
-  IntendToSwitchToSoleFinalOrder = '/intend-to-switch-to-sole-final-order.njk',
-  ServiceApplicationRejected = '/service-application-rejected.njk',
 }
 
 export const enum TranslatedToLanguage {
