+++ src/main/steps/applicant1/hub-page/sole/soleTemplateSelector.ts
@@ -1,8 +1,9 @@
 import dayjs from 'dayjs';
 
 import { CaseWithId } from '../../../../app/case/case';
-import { HubTemplate, State, YesOrNo } from '../../../../app/case/definition';
+import { State, YesOrNo } from '../../../../app/case/definition';
 import { StateSequence } from '../../../state-sequence';
+import { HubTemplate } from '../hubTemplates';
 
 export const getSoleHubTemplate = (
   displayState: StateSequence,
